# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AdvanceClockInput": {
        "additionalProperties": false,
        "properties": {
          "seconds": {
            "maximum": 2592000.0,
            "minimum": 0.0,
            "title": "Seconds",
            "type": "integer"
          }
        },
        "required": [
          "seconds"
        ],
        "title": "AdvanceClockInput",
        "type": "object"
      },
...
      "ClaimItemInput": {
        "additionalProperties": false,
        "properties": {
          "claimItemId": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Claimitemid",
            "type": "string"
          },
          "claimedAmountToman": {
            "minimum": 0.0,
            "title": "Claimedamounttoman",
            "type": "integer"
          },
          "claimedQuantity": {
            "default": 1,
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Claimedquantity",
            "type": "integer"
          },
          "evidenceIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Evidenceids",
            "type": "array"
          },
          "paidTo": {
            "default": "specialist",
            "enum": [
              "specialist",
              "third_party"
            ],
            "title": "Paidto",
            "type": "string"
          },
          "reason": {
            "maxLength": 600,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "suppliedBy": {
            "$ref": "#/components/schemas/Party",
            "default": "specialist"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/LineType",
            "default": "extra"
          }
        },
        "required": [
          "claimItemId",
          "title",
          "claimedAmountToman",
          "reason"
        ],
        "title": "ClaimItemInput",
        "type": "object"
      },
...
      "DemoStateOut": {
        "additionalProperties": false,
        "properties": {
          "aiMode": {
            "title": "Aimode",
            "type": "string"
          },
          "appMode": {
            "title": "Appmode",
            "type": "string"
          },
          "clockOffsetSeconds": {
            "title": "Clockoffsetseconds",
            "type": "number"
          },
          "demoReferenceReady": {
            "title": "Demoreferenceready",
            "type": "boolean"
          },
          "serverTime": {
            "format": "date-time",
            "title": "Servertime",
            "type": "string"
          }
        },
        "required": [
          "appMode",
          "aiMode",
          "serverTime",
          "clockOffsetSeconds",
          "demoReferenceReady"
        ],
        "title": "DemoStateOut",
        "type": "object"
      },
      "DisputeOut": {
        "additionalProperties": false,
        "properties": {
          "claimItems": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Claimitems",
            "type": "array"
          },
          "currentProposal": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentproposal"
          },
          "customerClosedStatementsAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customerclosedstatementsat"
          },
          "decision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "evidenceDeadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidencedeadline"
          },
          "evidenceRoundsUsed": {
            "title": "Evidenceroundsused",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "resolvedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolvedat"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "settlement": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settlement"
          },
          "specialistClosedStatementsAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialistclosedstatementsat"
          },
          "statementDeadline": {
            "format": "date-time",
            "title": "Statementdeadline",
            "type": "string"
          },
          "statements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Statements",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DisputeStatus"
          }
        },
        "required": [
          "id",
          "status",
          "revision",
          "claimItems",
          "statementDeadline",
          "evidenceDeadline",
          "customerClosedStatementsAt",
          "specialistClosedStatementsAt",
          "evidenceRoundsUsed",
          "resolvedAt",
          "statements",
          "currentProposal",
          "decision",
          "settlement"
        ],
        "title": "DisputeOut",
        "type": "object"
      },
      "DisputeStatus": {
        "enum": [
          "dispute_open",
          "reviewing",
          "needs_evidence",
          "awaiting_ai",
          "resolved_by_ai",
          "resolved_by_agreement"
        ],
        "title": "DisputeStatus",
        "type": "string"
      },
      "EvaluationOut": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "changeVerdicts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Changeverdicts",
            "type": "array"
          },
          "displayState": {
            "title": "Displaystate",
            "type": "string"
          },
          "hasInvoiceOverprice": {
            "title": "Hasinvoiceoverprice",
            "type": "boolean"
          },
          "hasUnjustifiedIncrease": {
            "title": "Hasunjustifiedincrease",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isAssessable": {
            "title": "Isassessable",
            "type": "boolean"
          },
          "modelLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modellabel"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "$ref": "#/components/schemas/EvaluationStatus"
          }
        },
        "required": [
          "id",
          "status",
          "attempt",
          "changeVerdicts",
          "hasUnjustifiedIncrease",
          "hasInvoiceOverprice",
          "isAssessable",
          "note",
          "modelLabel",
          "displayState"
        ],
        "title": "EvaluationOut",
        "type": "object"
      },
      "EvaluationStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "needs_review",
          "failed"
        ],
        "title": "EvaluationStatus",
        "type": "string"
      },
...
      "GatewayOutcomeInput": {
        "additionalProperties": false,
        "properties": {
          "outcome": {
            "enum": [
              "succeed",
              "fail"
            ],
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "GatewayOutcomeInput",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenDisputeInput": {
        "additionalProperties": false,
        "properties": {
          "claimItems": {
            "items": {
              "$ref": "#/components/schemas/ClaimItemInput"
            },
            "maxItems": 40,
            "minItems": 1,
            "title": "Claimitems",
            "type": "array"
          }
        },
        "required": [
          "claimItems"
        ],
        "title": "OpenDisputeInput",
        "type": "object"
      },
...
      "Party": {
        "enum": [
          "customer",
          "specialist"
        ],
        "title": "Party",
        "type": "string"
      },
...
      "ResolutionProposalInput": {
        "additionalProperties": false,
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineItem"
            },
            "maxItems": 40,
            "title": "Lines",
            "type": "array"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "ResolutionProposalInput",
        "type": "object"
      },
...
      "RevisionInput": {
        "additionalProperties": false,
        "properties": {
          "expectedRevision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedrevision"
          }
        },
        "title": "RevisionInput",
        "type": "object"
      },
...
      "StatementInput": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "evidenceIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Evidenceids",
            "type": "array"
          },
          "itemPositions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 40,
            "title": "Itempositions",
            "type": "array"
          }
        },
        "required": [
          "body"
        ],
        "title": "StatementInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API نمایشی «ترب تعمیر». دادهٔ نمونه، پرداخت شبیه‌سازی‌شده و پاسخ AI نمایشی در این نصب صریحاً برچسب‌گذاری می‌شوند.",
    "title": "ترب تعمیر — API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/demo/advance-clock": {
      "post": {
        "description": "Move the demo clock forward only, so deadlines can be reached without waiting.",
        "operationId": "advance_clock_api_demo_advance_clock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceClockInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance Clock",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/demo/evaluations/{request_id}": {
      "get": {
        "operationId": "evaluation_for_api_demo_evaluations__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluation For",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/demo/gateway-outcome": {
      "post": {
        "description": "Force the next simulated payment or transfer to fail, to exercise the retry path.",
        "operationId": "queue_gateway_outcome_api_demo_gateway_outcome_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayOutcomeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Queue Gateway Outcome Api Demo Gateway Outcome Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue Gateway Outcome",
        "tags": [
          "demo"
        ]
      }
    },
...
    "/api/dispute-proposals/{proposal_id}/approve": {
      "post": {
        "description": "The second approval applies the settlement and expires any pending ruling.",
        "operationId": "approve_resolution_api_dispute_proposals__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Resolution",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/adjudicate": {
      "post": {
        "description": "Run the binding ruling.\n\nAllowed once the statement window has closed or both sides declared they are finished.\nA valid ruling applies without asking either party again.",
        "operationId": "adjudicate_api_disputes__dispute_id__adjudicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adjudicate",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/close-statements": {
      "post": {
        "description": "When both sides declare they are finished, adjudication may start early.",
        "operationId": "close_statements_api_disputes__dispute_id__close_statements_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Statements",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/proposals": {
      "post": {
        "operationId": "propose_resolution_api_disputes__dispute_id__proposals_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionProposalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Resolution",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/statements": {
      "post": {
        "description": "Editing a statement expires the frozen input, so a late ruling cannot land.",
        "operationId": "submit_statement_api_disputes__dispute_id__statements_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatementInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Statement",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/api/requests/{request_id}/dispute": {
      "get": {
        "operationId": "get_dispute_api_requests__request_id__dispute_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DisputeOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Dispute Api Requests  Request Id  Dispute Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "description": "One process per case; re-submitting the same dispute never creates a second one.",
        "operationId": "open_dispute_api_requests__request_id__dispute_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
  }
}